Fixed asset import / preview creation
- Texture previews are now created with method SPEED instead of BALANCED (doesn't help too much).
- Previews and properties for original asset are NOT recreated anymore if existing previews and properties are found (can save tremendous amount of preview creation time, but may requires manual preview sometimes).
- Fixed AssetService removing AssetProperties when it shouldnt
- Fixed asset import conflict options in JS client
- Asset import can now skip import of assets for which tasks already exist automatically (maybe original assets should not even be shown in selection tree if they are already attached to a task, similar to new assets)
- After asset import, backup is now triggered to reduce need to recreate previews during testing
- Fixed various aspects surrounding asynchronous preview creations:
- 3D Previews are now created async in the same way texture previews are (fixed deadlock bug caused by PythonTCPSocket not being thread-safe)
- Backups now always wait until async preview creation is done to save properties correctly. The caller can decide if he wants the backup method to block until backup is done (needed for any caller that wants to make backup before critical change). This can cause random stuff to get blocked for a very long time if previews are still being created. The JS client should show ongoing preview creation.
- Using CompletableFuture.join instead of .get to block until done

+++ src/main/java/gmm/service/tasks/ModelTaskService.java
@@ -4,8 +4,10 @@ import java.io.FileInputStream;
 import java.io.IOException;
 import java.io.OutputStream;
 import java.io.UncheckedIOException;
+import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
+import java.util.Optional;
 import java.util.concurrent.CompletableFuture;
 
 import org.apache.commons.io.IOUtils;
@@ -45,20 +47,28 @@ public class ModelTaskService extends AssetTaskService<ModelProperties> {
 
 	@Override
 	public CompletableFuture<ModelProperties> recreatePreview(
-			Path sourceFile, Path previewFolder, AssetGroupType type, ModelProperties asset) {
+			Path sourceFile, Path previewFolder, AssetGroupType type, Optional<ModelProperties> assetOrNull) {
 		
 		fileService.createDirectory(previewFolder);
 		final Path target = getPreviewFilePath(previewFolder, type);
-		deletePreview(target);
-		final MeshData meshData = python.createPreview(sourceFile, target);
-		final Set<String> texturePaths = new HashSet<>(String.class, meshData.getTextures());
-		final Set<String> textureNames = new HashSet<>(String.class);
-		for(final String path : texturePaths) {
-			textureNames.add(Paths.get(path).getFileName().toString());
-		}
-		asset.setTextureNames(textureNames);
-		asset.setPolyCount(meshData.getPolygonCount());
-		return CompletableFuture.completedFuture(asset);
+		
+		return CompletableFuture.supplyAsync(() -> {
+			final boolean hasAsset = assetOrNull.isPresent();
+			final ModelProperties asset = assetOrNull.orElse(newPropertyInstance());
+			
+			if (!type.isOriginal() || !Files.exists(target) || !hasAsset) {
+				deletePreview(target);
+				final MeshData meshData = python.createPreview(sourceFile, target);
+				final Set<String> texturePaths = new HashSet<>(String.class, meshData.getTextures());
+				final Set<String> textureNames = new HashSet<>(String.class);
+				for(final String path : texturePaths) {
+					textureNames.add(Paths.get(path).getFileName().toString());
+				}
+				asset.setTextureNames(textureNames);
+				asset.setPolyCount(meshData.getPolygonCount());
+			}
+			return asset;
+		});
 	}
 	
 	@Override
@@ -78,6 +88,9 @@ public class ModelTaskService extends AssetTaskService<ModelProperties> {
 		return previewFolder.resolve(isOriginalString + ".json");
 	}
 	
+	/**
+	 * Deletes preview if it exist.
+	 */
 	private void deletePreview(Path previewFile) {
 		if(previewFile.toFile().exists()) {
 			fileService.delete(previewFile);
